move coin icons to shared
